pyproject.toml
[project]
name = "jee25-solutions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "dspy-ai>=2.6.24",
    "google-genai>=1.17.0",
    "jupyterlab>=4.4.3",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]
